Add direct connection end reason
